Split the card system out of graph-item-content, and test its rules
graph-item-content.tsx was 2,273 lines holding two card bodies, the drag
ghost, both trim slots, eight store subscriptions and every chip and
placeholder they share. It is 163 lines now: the shell that decides which
card kind renders, and the CollectionsComponents table.

The point of the split is coverage, not tidiness. This app's vitest cannot
parse .tsx, so a rule written inside a card component is untestable by
construction. Six of them move to graph-card-model.ts and gain 29 tests:
enabledChildCount, firstChildIsAudio, cardDimming, cardVideoFrameCount,
collectionCardItemCount and collectionCardSeconds.

Extracting the dimming rule found the thing that makes it worth doing. The
precedence ladder (drag source > disabled > filter miss) was written TWICE,
once per card kind, each under a comment promising it was "shared so the two
kinds cannot drift apart". They were not shared, they were copied. One
function decides it now and one map in graph-card-dimming.ts turns it into
classes — inside the Tailwind-scanned tree, because a class name in lib/ is
never generated and fails silently as unstyled pixels.

Behaviour is unchanged. Every comment travels with the code it explains.

CARD_CONTROL_INSET_RIGHT/_TOP turn out to have no importers at all —
CardCornerSlot carries a literal twin — so they are moved rather than
deleted; that is a separate decision.

Verified: tsc clean, 864 app tests (was 835), lint clean with the same 5
pre-existing <img> warnings, graph-view e2e 169/169.
