+++ apps/timeline-gstudio001/components/graph-view/graph-card-trim.tsx
@@ -1,0 +1,93 @@
+"use client";
+
+import { memo } from "react";
+
+import type {
+  CollectionTrimHandleContentProps,
+  CollectionTrimOverviewContentProps,
+} from "@storyboard/ui/dnd-collections";
+
+import { OVERVIEW_FRAME_SIZE, overviewFrameCount } from "@/lib/card-ghost-frames";
+import { videoFrameUrls } from "@/lib/video-frame-url";
+
+/**
+ * The trim overview's background (the "sequence above" a selected video),
+ * replacing the package default via the `OverviewContent` registry slot. The
+ * default TILES the 1–2 stored posters by modulo — with one poster every
+ * frame is the same still, which reads as broken (R7 #2). This samples each
+ * slot at its own time across the FULL SOURCE through the same
+ * `videoFrameUrls` seam the card filmstrip uses, so the two strips show the
+ * same kind of sequence — including the last-slot end-frame pin (R7 #3).
+ * The default's "full clip x.xs" readout is dropped on request (R7 #3); the
+ * trim readouts around the window already carry the numbers.
+ *
+ * Memo comparator (mirrors the package default's): the contract carries live
+ * trimIn/trimOut that change per pointer move during a drag, but this
+ * renderer reads only `node` and `fullWidth` — shallow memo would reconcile
+ * up to 40 <img> elements per move for nothing.
+ */
+export const GraphTrimOverviewContent = memo(
+  function GraphTrimOverviewContent({ node, fullWidth }: CollectionTrimOverviewContentProps) {
+    const posters = node.posterSrcs ?? [];
+    // Enough square frames to cover the strip width (ceil so the row fills;
+    // the container clips the overflow), capped so a long source stays
+    // bounded.
+    const frameCount = overviewFrameCount(fullWidth);
+    // The overview shows the FULL source (the amber window marks the visible
+    // part), so the sample range is [0, fullDurationSeconds] — not the
+    // trimmed range the card uses.
+    const frameSrcs = videoFrameUrls(posters, frameCount, {
+      trimInSeconds: 0,
+      effectiveSeconds: Math.max(0, node.fullDurationSeconds),
+    });
+    return (
+      <div className="flex h-full w-full">
+        {frameSrcs.length === 0 ? (
+          <span className="flex h-full w-full items-center justify-center bg-muted text-[11px] text-muted-foreground select-none">
+            No preview frames
+          </span>
+        ) : (
+          frameSrcs.map((src, index) => (
+            // eslint-disable-next-line @next/next/no-img-element
+            <img
+              key={index}
+              src={src}
+              alt=""
+              draggable={false}
+              style={{ width: OVERVIEW_FRAME_SIZE }}
+              className="h-full shrink-0 border-r border-black/60 object-cover last:border-r-0"
+            />
+          ))
+        )}
+      </div>
+    );
+  },
+  (prev, next) =>
+    prev.node === next.node &&
+    prev.pixelsPerSecond === next.pixelsPerSecond &&
+    prev.fullWidth === next.fullWidth,
+);
+
+export const GraphTrimHandle = memo(function GraphTrimHandle({
+  side,
+}: CollectionTrimHandleContentProps) {
+  // Handles exist only on SELECTED clips (trimRequiresSelection at the
+  // provider), so these pixels are always the active affordance — no
+  // hover-reveal state for unselected cards to style anymore.
+  return (
+    <span
+      className={[
+        // blue-500 — the SELECTION colour (the ring on a selected card, the
+        // count in the select row). These handles only exist on a selected
+        // clip, so wearing the selection's colour is what ties them to the
+        // thing they act on. Amber is left over from when amber WAS the
+        // selection colour; once selection moved to blue it read as a second,
+        // unrelated accent on the one card already wearing the first.
+        "flex h-full w-full items-center justify-center bg-blue-500 opacity-95",
+        side === "left" ? "rounded-l-md" : "rounded-r-md",
+      ].join(" ")}
+    >
+      <span className="h-4 w-0.5 rounded bg-black/60" />
+    </span>
+  );
+});
